Create an authentication rspec helper for logging in
