+++ spec/requests/graphql/mutation/stripe_checkout_session_create_spec.rb
@@ -63,34 +63,7 @@ RSpec.describe "StripeCheckoutSessionCreate", type: :request do
 
           user.update(stripe_customer_id: customer.id)
 
-          # Sign in user - convert this into a helper
-          sign_in_query =
-            <<-GRAPHQL
-                mutation CreateSession($input: SessionCreateInput!) {
-                    sessionCreate(input: $input){
-                        user {
-                            id
-                        }
-                        errors {
-                            message
-                            path
-                        }
-                    }
-                }
-            GRAPHQL
-
-          sign_in_variables = {
-            input: {
-                sessionInput: {
-                    email: user.email,
-                    password: user.password
-                }
-            }
-          }
-
-          post "/graphql", params: { query: sign_in_query, variables: sign_in_variables }
-
-          expect(controller.current_user).to be_present
+          login_user(user)
 
           product_list = Stripe::Product.list
           line_items = product_list.data.map { |product| { price: product.default_price, quantity: 1 } }
